Added getNodes() on HM so you can pass a bunch of entities to convert them, that way methods that return a bunch of entities don't have to convert them, thereby filling the memory, unless the user decides it's necessary.
Added getId() so we can actually read the Id value of the root node otherwise we couldn't
know which children belong to which root.

Added default values to the configuration class so that you can just use it as is if you
follow the convention and don't need to set up your own config class.

Signed-off-by: Jordi Boggiano <j.boggiano@seld.be>